Pull common members and setup to base class
